+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is used for recording Invader's changes. This changelog is based on [Keep a
 - invader-bitmap: Double multiply sprites now simply replaces the pixel like tool.exe rather than alpha blend into gray
 - invader-bitmap: Usage and the p8 compression flag are now preserved
 - invader-bitmap: Height maps now generate bump maps similar to tool.exe
+- invader-bitmap: Added --usage (default or bumpmap) and --bump-height parameters
 
 ### Fixed
 - invader-bitmap: Fixed some issues with spacing